eblnn/requirements.txt
torch>=2.0
ncps>=0.0.6
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
wandb>=0.16
